move multireader out of /pkg
