+++ daemon/logger/jsonfilelog/multireader/multireader_test.go
@@ -1,4 +1,4 @@
-package ioutils
+package multireader
 
 import (
 	"bytes"
